requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
openai==1.93.0
pinecone==7.3.0
pymongo==4.13.2
python-dotenv==1.1.1